Cart_add form, view
in views.py add_cart form has advanced functionality. First, it checks whether the user has his own cart or not. If not, creates it. Then checks if there is an item with similar name in cart. If yes, adds additional quantity to it. Else, creates a new CartItem object.